+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # renv 0.14.0 (UNRELEASED)
 
+* Fixed an issue where `renv` would fail to load if the `~/.Rprofile` existed
+  but was an empty file.
+
 # renv 0.13.0
 
 * `renv::snapshot()` no longer creates an `renv/activate.R` file in the project